Pagination.

+++ src/api/repositories/MoviesPostgresRepository.ts
@@ -4,6 +4,11 @@ import { MoviesRepository } from "./MoviesRepository";
 import { connection } from "../database/connection";
 import { Client } from "pg";
 
+export interface GetAllResponse {
+  data: Movie[];
+  count: number;
+}
+
 class MoviesPostgresRepository implements MoviesRepository {
   private client: Client;
 
@@ -28,27 +33,38 @@ class MoviesPostgresRepository implements MoviesRepository {
     );
   }
 
-  async getAll(): Promise<Movie[]> {
-    const response = await this.client.query(
-      "SELECT * FROM MOVIES;"
-    );
-
-    return response.rows;
-  }
-
   async getById(id: string): Promise<Movie> {
     const response = await this.client.query(
       "SELECT * FROM MOVIES WHERE MOVIES.id = $1;", [id]
     );
 
-    console.log(response);
     return response.rows[0];
   }
 
-  async delete(id: string): Promise<void> {
-    await this.client.query(
-      "DELETE FROM MOVIES WHERE MOVIES.id = $1;", [id]
-    );
+  async getAll(startIndex: number, endIndex: number): Promise<GetAllResponse> {
+    const response1 = this.client.query("SELECT * FROM MOVIES LIMIT $1 OFFSET $2;", [endIndex, startIndex]);
+    const response2 = this.client.query("SELECT COUNT(*) FROM MOVIES;");
+
+    const promises = [response1, response2];
+
+    const response = await Promise.all(promises);
+
+    return {
+      data: response[0].rows,
+      count: response[1].rows[0].count,
+    }
+  }
+
+  async delete(id: Movie["id"]): Promise<number> {
+
+    if (await this.getById(id)) {
+      await this.client.query(
+        "DELETE FROM MOVIES WHERE MOVIES.id = $1;", [id]
+      );
+      return 200;
+    } else {
+      return 404;
+    }
   }
 }
